refactor: Update authentication middleware to return error messages instead of redirecting

+++ src/routes/apiCarts.router.js
@@ -1,9 +1,10 @@
 import { Router } from "express";
 import cartService from "../services/cartService.js";
-
+import auth from "../middlewares/auth.js";
+import isAdmin from "../middlewares/isAdmin.js";
 const router = Router();
 
-router.post('/',  async (req, res) => {
+router.post('/', auth, async (req, res) => {
   try {
     const userId = req.session.user._id
     const cart = await cartService.addCart(userId);
@@ -13,7 +14,7 @@ router.post('/',  async (req, res) => {
   }
 })
 
-router.get('/:cid', async (req, res) => {
+router.get('/:cid', auth, isAdmin, async (req, res) => {
   const cartId = req.params.cid;
   try {
     const cart = await cartService.getCart(cartId);
@@ -23,7 +24,7 @@ router.get('/:cid', async (req, res) => {
   }
 })
 
-router.get('/',  async (req, res) => {
+router.get('/', auth, isAdmin, async (req, res) => {
   try {
     const carts = await cartService.getAllCarts();
     res.status(200).send({status:'success', message:'carritos encontrados', carts});
@@ -32,9 +33,15 @@ router.get('/',  async (req, res) => {
   }
 })
 
-router.post('/:cid/products/:pid', async (req, res) => {
+router.post('/:cid/products/:pid', auth, async (req, res) => {
   const cartId = req.params.cid;
   const productId = req.params.pid;
+  const userId = req.session.user._id;
+
+  const cart = await cartService.getCart(cartId);
+  if (cart.user !== userId) {
+    return res.status(401).send({status:'error', message:'No tienes permisos para agregar productos a este carrito'})
+  }
 
   const quantity = +req.body.quantity;
   if (quantity) {
@@ -52,10 +59,16 @@ router.post('/:cid/products/:pid', async (req, res) => {
   }
 })
 
-router.put('/:cid/products/:pid', async (req, res) => {
+router.put('/:cid/products/:pid', auth, async (req, res) => {
   const cartId = req.params.cid
   const productId = req.params.pid
   const quantity = +req.body.quantity;
+  const userId = req.session.user._id;
+
+  const cart = await cartService.getCart(cartId);
+  if (cart.user !== userId) {
+    return res.status(401).send({status:'error', message:'No tienes permisos para modificar este carrito'})
+  }
   
   if (!quantity || typeof quantity !== 'number' || quantity <= 0) {
     return res.status(400).send({ status: 'error', error: 'Invalid quantity' });
@@ -69,8 +82,14 @@ router.put('/:cid/products/:pid', async (req, res) => {
   }
 })
 
-router.delete("/:cid", async (req, res) => {
+router.delete("/:cid", auth, async (req, res) => {
   const cartId = req.params.cid;
+  const userId = req.session.user._id;
+
+  const cart = await cartService.getCart(cartId);
+  if (cart.user !== userId) {
+    return res.status(401).send({status:'error', message:'No tienes permisos para eliminar este carrito'})
+  }
 
   try {
     const cart = await cartService.deleteAllProductsFromCart(cartId);
@@ -80,9 +99,16 @@ router.delete("/:cid", async (req, res) => {
   }
 })
 
-router.delete("/:cid/products/:pid", async (req, res) => {
+router.delete("/:cid/products/:pid", auth, async (req, res) => {
   const cartId = req.params.cid;
   const productId = req.params.pid
+  const userId = req.session.user._id;
+
+  const cart = await cartService.getCart(cartId);
+  if (cart.user !== userId) {
+    return res.status(401).send({status:'error', message:'No tienes permisos para eliminar productos de este carrito'})
+  }
+  
   try {
     const cart = await cartService.deleteProductFromCart(cartId, productId)
     res.status(200).send({status:'success', message:`producto ${productId} eliminado del carrito`, cart});
